Rebase master (#2)
* Add startWith and startAt to TapTargetSequence (#228)

Previously it was cumbersome to restore state in a TapTargetSequence
after rotation due to the fact that you needed to manually modify the
list of targets you pass in to get back to your desired target. Now the
class supports starting from a specified target id or position in the
queue, allowing callers to keep a static list of targets for their
sequence.

* Added Support for React Native Projects (#230)

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Add notion of isDismissing and gate onReady() callbacks on it (#243)

* Add notion of isDismissing and gate onReady() callbacks on it

There was a race condition I came across where a view could be animating out at the same time as the animation for dismissing the tap target. If the view animated out before the taptarget, it would trigger a layout again, and potentially redisplay the taptarget again. This gates onReady callbacks against the notion of "isDismissing". Tested locally in my project and confirmed this resolves the issue.

* Return early (and earlier!)

* Fix NPE when dismiss is called immediately (#244)

Small "race" condition that shouldn't crash. Fixes #240

* 1.11.0

* Fix "sometimes" typo (#246)

* Update Gradle version (#253)

* Disable interaction during expand animation (#248)

* Fix build (#256)

* Update to latest support library + SDK version (#257)

Was pretty outdated previously.

* Use api configuration for public dependencies. (#258)

Some support annotations and the v7 support Toolbar are both part of the
public api, so they need to be specified as `api` in order for them to
transitively picked up by downstream dependants.

* Update build tools to most recent ones (#261)

* update build tools to most recent ones

* revert to last stable version and use gradle 4.5

* Use central build variables (#264)

* in module "taptargetview" buildToolsVersion was missing und request an install of 26.0.2
* variable SupportLibraryVersion was just in one module

* Handle dismiss before show

* Replacement of "compile" with "implementation" (#291)

As "compile" is going to be deprecated end-2018, this PR replaces with "implementation" that is the new usage.

* Fix NPE from dismissal before determining center (#298)

Fixes #279

* Fix not being able to cancel on last target (#299)

Simplifies an early return + properly unsets currentView. Fixes #294

* Fix decor related issues (#300)

This fixes issues relating to transparent nav bars, status bars and
multi window mode. The library will now detect when the window it is
attached to has a transparent system bar and properly adjust its clip
bounds. Fixes #233

* Release 1.12.0

* Updated readme to reflect latest version